+++ mayan/apps/logging/classes.py
@@ -1,7 +1,7 @@
 import logging
 
 from django.apps import apps
-from django.db.models.signals import pre_delete
+from django.db.models.signals import post_save, pre_delete
 
 from mayan.apps.acls.classes import ModelPermission
 from mayan.apps.common.menus import menu_list_facet
@@ -15,6 +15,10 @@ logger = logging.getLogger(name=__name__)
 
 class ErrorLog:
     _registry = {}
+
+    @staticmethod
+    def get_model_instance_partition_name(model_instance):
+        return '{}.{}'.format(model_instance._meta.label, model_instance.pk)
 
     @classmethod
     def all(cls):
@@ -34,26 +38,14 @@ class ErrorLog:
         return str(self.app_config.verbose_name)
 
     def register_model(self, model, register_permission=False):
-        ErrorLogModel = apps.get_model(
-            app_label='logging', model_name='ErrorLog'
-        )
-        try:
-            self.model, created = ErrorLogModel.objects.get_or_create(
-                name=self.app_config.name
-            )
-        except ErrorLogModel.MultipleObjectsReturned:
-            # Self heal previously repeated entries.
-            ErrorLogModel.objects.filter(name=self.app_config.name).delete()
-            self.model, created = ErrorLogModel.objects.get_or_create(
-                name=self.app_config.name
-            )
-
         error_log_instance = self
 
         @property
         def method_instance_logs(self):
-            error_log_partition, created = error_log_instance.model.partitions.get_or_create(
-                name='{}.{}'.format(model._meta.label, self.pk)
+            error_log_partition = error_log_instance.stored_error_log.partitions.get(
+                name=ErrorLog.get_model_instance_partition_name(
+                    model_instance=self
+                )
             )
 
             error_log_partition.entries.exclude(
@@ -73,11 +65,46 @@ class ErrorLog:
                 model=model, permissions=(permission_error_log_view,)
             )
 
-        def handler_delete_model_entries(sender, instance, **kwargs):
-            instance.error_log.all().delete()
+        def handler_model_instance_delete_partition(sender, instance, **kwargs):
+            return self.stored_error_log.partitions.get(
+                name=ErrorLog.get_model_instance_partition_name(
+                    model_instance=instance
+                )
+            ).delete()
 
-        pre_delete.connect(
-            dispatch_uid='logging_handler_delete_model_entries',
-            receiver=handler_delete_model_entries,
+        def handler_model_instance_create_partition(sender, instance, **kwargs):
+            if kwargs['created']:
+                return self.stored_error_log.partitions.create(
+                    name=ErrorLog.get_model_instance_partition_name(
+                        model_instance=instance
+                    )
+                )
+
+        post_save.connect(
+            dispatch_uid='logging_handler_model_instance_create_partition',
+            receiver=handler_model_instance_create_partition,
             sender=model, weak=False
         )
+        pre_delete.connect(
+            dispatch_uid='logging_handler_model_instance_delete_partition',
+            receiver=handler_model_instance_delete_partition,
+            sender=model, weak=False
+        )
+
+    @property
+    def stored_error_log(self):
+        StoredErrorLog = apps.get_model(
+            app_label='logging', model_name='StoredErrorLog'
+        )
+        try:
+            stored_error_log, created = StoredErrorLog.objects.get_or_create(
+                name=self.app_config.name
+            )
+        except StoredErrorLog.MultipleObjectsReturned:
+            # Self heal previously repeated entries.
+            StoredErrorLog.objects.filter(name=self.app_config.name).delete()
+            stored_error_log, created = StoredErrorLog.objects.get_or_create(
+                name=self.app_config.name
+            )
+
+        return stored_error_log