+++ mayan/apps/logging/tests/mixins.py
@@ -9,17 +9,17 @@ from .literals import TEST_ERROR_LOG_ENTRY_RESULT
 
 class ErrorLoggingTextMixin:
     def _create_error_log_test_object(self):
-        self._create_test_user()
         self.test_model = get_user_model()
-        self.test_object = self.test_user
-
-    def _create_error_log_entry(self):
         app_config = apps.get_app_config(app_label='logging')
         self.error_log = ErrorLog(app_config=app_config)
         self.error_log.register_model(
             model=self.test_model, register_permission=True
         )
 
+        self._create_test_user()
+        self.test_object = self.test_user
+
+    def _create_error_log_entry(self):
         self.error_log_entry = self.test_object.error_log.create(
             text=TEST_ERROR_LOG_ENTRY_RESULT
         )